Add comprehensive testing framework and configuration for LMSRMarket
- Introduced demo scripts for running comprehensive tests directly with Hardhat.
- Created a new Hardhat configuration file to support the updated testing framework.
- Added MockERC20 contract for simulating USDC transactions in tests.
- Enhanced test coverage with detailed scenarios for market creation, share buying, and resolution.
- Updated package dependencies for improved compatibility and functionality.
- Documented the success of the testing framework in a new SUCCESS_SUMMARY.md file.

+++ contracts/ABDKMath64x64.sol
@@ -21,12 +21,12 @@ library ABDKMath64x64 {
 
     function fromUInt(uint256 x) internal pure returns (int128) {
         require(x <= 0x7FFFFFFFFFFFFFFF);
-        return int128(int256(x << 64));
+        return int128(int256(x) << 64);
     }
 
-    function toUInt(int128 x) internal pure returns (uint64) {
+    function toUInt(int128 x) internal pure returns (uint256) {
         require(x >= 0);
-        return uint64(uint128(x >> 64));
+        return uint256(uint128(x >> 64));
     }
 
     function add(int128 x, int128 y) internal pure returns (int128) {
